Update nixpkgs.
This updates GHC from 8.10.7 to 9.0.2.
A small fix was necessary due to the new GHC version.

We now follow the nixpkgs branch/channel `master`.
The reason for this is, that `nixpkgs-unstable` and `nixos-unstable` are
literally just retarded versions of `master`.
For development purpose it makes sense to stay up-to-date.
Deployments should probably stick to more a more stable nixpkgs version.